+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Version 0.5.0
 
 - Pass multiple traces to efel at the same time (see NSETM-1780)
 - Add CV-validation (supports both voltage and current clamp modes)
+- Fix compatibility with BGLibPy>=4.7.2.
 
 Version 0.4.1
 -------------